requirements.txt
streamlit
pandas
numpy
scikit-learn
groq
python-dotenv
